start thumb decoding, switch to c++
started on thumb decoding. from looking at the instruction table, only the upper 10-bits are needed for decoding. because of this, i have shifted and masked the opcode, this reduces the number of cases in the jump table generated by compiler*

(NOTE: i haven't checked if the compiler [gcc/clang] actually generates a jump table or a if else chain).

I have decided to use c++14 for now as it has official support for binary literals (0b) and spacers in numbers (0b000'0000'0000). Still going to keep it C-style and only use std::array from stl as i don't want slow compile time from stl bloat.

The thumb decoding switch seems to be correct so far, though i still need to:
- check all instructions.
- check each instruction encoded value.