start work on barrel shifter + arm data processing
i decided to make as many functions pure as possible. reason for this is that it allows the functions to be easily tested at compile-time, so if it fails to compile, either my test is wrong or the tested function is.

this really helps speedup testing functions as i go along, as i can just write the function, then directly under it call it in static_assert, and without needing to manually compile, vscode will error if it assert fails.