start to add memory interface
this went through a few designs...

first, i created a base class with pure virtual methods for read / write. then i created 2 inherited classes called static and dynamic.

static was to be used for constexpr, dynamic was for runtime (until we get constexpr vector)

the arm7tdmi struct would now take a template arg for the memory interface, in which you could pass either static or dynamic classes.

however, i found out that there was no need for dynamic memory as the gba has a max* rom size of 32mb (usually) and max sram size of 64kb.

because of this, the data could be normal std::array, and thus the base class stuff was scrapped.

i still liked the idea of passing the memory interface as a template arg, as it was like connecting different components are compile time, rather than attaching them at runtime with refrences / shared_ptr.

however this caused issues for all my current arm / thumb functions, as they are not methods as want the arm7tdmi struct as an arg, id need to add a template to every single function.

the reason these functions are seperate and are not methods is because theres no clean way to write it otherwise. all methods (to be constexpr) would need to be defined in the same class header, which is *a lot* of instructions and would easily become a 4k LOC header which is too big to manage.

so i decided to split the functions into groups and placed them in different headers, making it much easier to focus on what i need to work on. this means explicitly passing in the arm7tdmi struct, rather than implicit *this*.